refactor(server): extract reducers to directory and types module
