+++ server/spacetimedb/Cargo.toml
@@ -3,10 +3,8 @@ name = "iron_admiral"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 spacetimedb = { version = "2.0.2" }
