+++ server/spacetimedb/src/lib.rs
@@ -1,3 +1,10 @@
 pub mod physics;
 pub mod reducers;
 pub mod tables;
+pub mod types;
+
+pub use physics::*;
+pub use tables::{PhysicsTimer, Player, Ship};
+pub use types::Waypoint;
+// Reducers are usually accessed by name in SpacetimeDB, so we can export them individually if needed
+// or just leave them in their modules.
