+++ server/spacetimedb/src/types.rs
@@ -1,0 +1,8 @@
+use spacetimedb::SpacetimeType;
+
+#[derive(SpacetimeType, Clone, Debug, PartialEq)]
+pub struct Waypoint {
+    pub x: f32,
+    pub y: f32,
+    pub target_speed: f32,
+}
